feat: Add RegistrationCode struct
Add two structs to the Registration namespace. The first, RegistrationCodeHelper, is all constants and static functions to help check and generate only A-Z and 0-1.

The second struct is the actual RegistrationCode, with constructors to initialize from int[], char[] or string. It also has a function to generate a new code from a random instance.